+++ tests/lantern_tests/e2e/test_site_e2e.py
@@ -20,3 +20,17 @@ class TestSentry:
 
         page.locator("text=Is something wrong with this page?").click()
         expect(page.locator("text=Add a screenshot")).to_be_visible()
+
+    def test_fallback(self, fx_exporter_static_server: Popen, page: Page):
+        """Email link shown where JavaScript is unavailable."""
+        context = page.context.browser.new_context(java_script_enabled=False)
+        no_js_page = context.new_page()
+        no_js_page.goto("http://localhost:8123/legal/privacy/index.html")
+        status_code = no_js_page.evaluate("window.performance.getEntries()[0].responseStatus")
+        assert status_code == 200
+
+        fallback_link = no_js_page.get_by_role("link", name="Is something wrong with this page?")
+        expect(fallback_link).to_be_visible()
+        href = fallback_link.get_attribute("href")
+        assert href.startswith("mailto:")
+        context.close()